chore: move integration tests to dedicated 'integration' directory
