+++ src/test/integration/composite-escape.test.ts
@@ -11,7 +11,7 @@ import {
     sendEscapeKey,
     openTextDocument,
     sendInsertKey,
-} from "../integrationUtils";
+} from "./integrationUtils";
 
 describe("Composite escape key", () => {
     let client: NeovimClient;
