+++ src/test/integration/multi-byte-chars.test.ts
@@ -10,7 +10,7 @@ import {
     sendEscapeKey,
     openTextDocument,
     sendInsertKey,
-} from "../integrationUtils";
+} from "./integrationUtils";
 
 describe("Multi-width characters", () => {
     let client: NeovimClient;
